requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
jinja2==3.1.4
starlette==0.38.5
# PDF processing
# PyMuPDF==1.23.14  # Removed: Not used in code and incompatible with Python 3.13 on Render
PyPDF2>=3.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-multipart>=0.0.9
# Vector store and embeddings
faiss-cpu>=1.8.0
openai==1.12.0
groq==0.28.0
google-generativeai==0.3.2
# Text processing
nltk==3.8.1
# sentence-transformers==2.2.2  # Optional: Only needed if no OpenAI API key. Comment out if build fails on Render
# File watching
watchdog==4.0.0
# Task scheduling
apscheduler==3.10.4
python-dateutil==2.8.2




